+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog for package bob_llm
 ------------------
 * Added support for loading system_prompt from files and new system_prompt_file parameter
 * Implemented dynamic parameter reconfiguration for LLM client and system prompt
+* Added optional `eof` parameter to signal the end of a stream on `llm_stream`
 * Removed prefix v1 from chat API path
 
 1.0.2 (2026-02-01)